Swap i and nbits in Uintn/Intn methods
